refactor: move path utils out of viewer file
